:hammer: removing useless cast

+++ src/game_entity/Hitbox.java
@@ -140,10 +140,10 @@ public class Hitbox {
     public void draw(Graphics2D g2d, int screenX, int screenY, Color color) {
         g2d.setColor(color);
         g2d.drawRect(
-                (int) (screenX),
-                (int) (screenY),
-                (int) Constants.TILE_SIZE,
-                (int) Constants.TILE_SIZE
+                (screenX),
+                (screenY),
+                Constants.TILE_SIZE,
+                Constants.TILE_SIZE
         );
     }
